requirements.txt
certifi==2024.8.30
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
psycopg2-binary==2.9.9
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
shapely==2.0.6
six==1.16.0
tzdata==2024.2
pydantic~=2.9.2
celery~=5.4.0
SQLAlchemy~=2.0.39
fastapi~=0.115.11
uvicorn~=0.34.0